feat(stella-protocol,stella-core,stella-tui): typed TurnParked/TurnWoken wire events for parked waits
The parked-wait primitive (#1471) announced the park and the wake as
synthetic AgentEvent::Text deltas, so consumers could not tell a park
from model narration and replay could not attribute the wall-clock gap.

- AgentEvent::TurnParked { description, poll_interval_secs,
  deadline_secs } / TurnWoken { reason, polls_used } beside Steered,
  with wire tags turn_parked/turn_woken and serde round-trip witnesses
  (invariant 4). event.rs's inline test module moves to
  src/event/tests.rs (the anthropic/tests.rs pattern) so the wire
  contract stays under its file-size ceiling.
- Bus lifecycle names agent.turn.parked / agent.turn.woken
  (observer-only; names::BLOCKING untouched). The names catalog moves
  to bus/names.rs (the driver/settlement.rs pattern) for the same
  ceiling reason. Payloads carry the timing envelope and the WakeReason
  token only — never the tool-authored description (content-free rule).
- driver/waiting.rs emits the typed pair and drops the ⏳/▶ prose
  deltas; WakeReason::wire_token() pins the reason spelling.
- The deck folds the pair into typed Parked/Woken transcript entries,
  the plain surface words them in textline (parked/woken), the fleet
  deck traces them cool (intensity 60) and keeps the lane Running.
- Pipeline replay: signatures for totality; excluded from
  structural_diff's positional walk (external-state timing, additive
  vs old goldens — the SpeculationDiscarded + SubAgent grounds).
- The parked_wait witnesses now assert the typed events, including the
  deadline path's polls accounting.

Refs #1471
Closes #1857
